refactor(phase8-b): move context/source/query to canonical locations (D5)
Per 符炫炜 D5 canonical (msg=094f9bee → architecture v0):
- aperag/context/ → aperag/domains/retrieval/context/ (retrieval-only consumer)
- aperag/source/ → aperag/platform/source/ (cross-domain shared infra)
- aperag/query/ → aperag/platform/query/ (cross-domain shared infra; DocumentWithScore primitive)

Mechanical changes:
- git mv preserves history (rename detection)
- 15 import sites updated across aperag/ + tests/
- 0 hard-coded path references remain (grep verified)
- Empty top-level dirs (aperag/context, aperag/source, aperag/query) removed
- aperag/platform/__init__.py created (new namespace; #38/#39 will add their own subtrees)

Scope guard (per PM msg=c9ff2c19 + msg=e9170089):
- Did NOT touch aperag/chat/ aperag/graph_curation/ (out of #37 scope)
- Did NOT touch aperag/concurrent_control/ (#38 owner)
- Did NOT touch aperag/db/ aperag/auth/ aperag/views/auth.py (#36/#39 owners)
- celery domainization in subsequent commit (D4 part of #37)

Ghost-check: none.
