+++ aperag/platform/source/upload.py
@@ -17,8 +17,8 @@ from typing import Any, Dict, Iterator
 
 from aperag.objectstore.base import get_object_store
 from aperag.schema.common import CollectionConfig
-from aperag.source.base import LocalDocument, RemoteDocument, Source
-from aperag.source.utils import gen_temporary_file
+from aperag.platform.source.base import LocalDocument, RemoteDocument, Source
+from aperag.platform.source.utils import gen_temporary_file
 
 
 class UploadSource(Source):
