+++ aperag/utils/weixin/client.py
@@ -20,7 +20,7 @@ from typing import Any, Dict
 
 import aiohttp
 
-from aperag.source.base import CustomSourceInitializationError
+from aperag.platform.source.base import CustomSourceInitializationError
 from aperag.utils.history import get_async_redis_client
 
 logger = logging.getLogger(__name__)
