+++ tests/unit_test/test_es_p0_contract.py
@@ -8,7 +8,7 @@ from aperag.domains.indexing.db.models import DocumentIndexType
 from aperag.domains.knowledge_base.db.models import CollectionStatus
 from aperag.domains.knowledge_base.service.document_service import DocumentService
 from aperag.domains.retrieval.pipeline import SearchPipelineService
-from aperag.query.query import DocumentWithScore
+from aperag.platform.query.query import DocumentWithScore
 from aperag.tasks.collection import CollectionTask
 from aperag.tasks.document import DocumentIndexTask
